+++ src/components/navbar/info-bar.tsx
@@ -154,7 +154,7 @@ export const InfoBar = ({ scrollY }: { scrollY: number }) => {
                   </AvatarFallback>
                 </Avatar>
               </DropdownMenuTrigger>
-              <DropdownMenuContent className="w-56">
+              <DropdownMenuContent className="relative w-56">
                 <DropdownMenuLabel>My Account</DropdownMenuLabel>
                 <DropdownMenuSeparator />
                 <DropdownMenuGroup>
